requirements.txt
Flask==1.0.2
PyJWT==1.6.4
pylint==1.9.3
nose==1.3.7
pinocchio==0.4.2
coverage==4.5.2
